test(web): split score analytics server tests
Split the slow score comparison analytics suite into parallelizable files and reduce the sampling fixture size while preserving branch coverage.
